refactor: Integrate modal and alert using ModalProvider and AlertProvider in layout
