+++ src/app/(before-login)/(without-navbar)/layout.tsx
@@ -1,27 +1,14 @@
-"use client";
-
 import { ReactNode } from "react";
-import { useAlertStore } from "@/lib/hooks/useAlertStore";
-import Alert from "@/components/common/alert/Alert";
-import { useRouter } from "next/navigation";
+import AlertProvider from "@/components/common/alert/AlertProvider";
 
 export default function Layout({ children }: { children: ReactNode }) {
-  const { currentAlert } = useAlertStore();
-  const router = useRouter();
-
   return (
     <div className="flex min-h-screen">
       <div className="flex flex-col">
         {/* page content */}
         {children}
       </div>
-
-      {/* Alert */}
-      {currentAlert === "passwordMismatch" && <Alert />}
-      {currentAlert === "emailDuplicated" && <Alert />}
-      {currentAlert === "signupSuccess" && (
-        <Alert onConfirm={() => router.push("login")} />
-      )}
+      <AlertProvider />
     </div>
   );
 }
