+++ src/lib/hooks/useGetViewport.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useState, useEffect } from "react";
 import debounce from "lodash.debounce";
